Fix crash when Curios isn't installed
Fix a crash when Curious isn't installed.

Prepare for release.
