+++ gradle.properties
@@ -9,7 +9,7 @@ neo_version=26.1.2.44-beta
 neo_version_range=[26.1.2.33-beta,)
 loader_version_range=[4.0,)
 mod_id=iwcompatbridge
-mod_version=26.1.2-1.14.0
+mod_version=26.1.2-1.14.1
 mod_group_id=tech.anonymoushacker1279.iwcompatbridge
 immersiveweapons_version=26.1.2-1.37.0
 openportals_version=26.1.2-1.2.0
